Implemented ViewFinderTrait
